:recycle: (SocialLogin) 확장성을 위해 역할을 분리하였습니다.
